backend/requirements.txt
Flask
Flask-Cors
woocommerce
python-dotenv
requests
gunicorn==21.2.0 